Update train_controlnet.py
reduce float value error for bfloat16

+++ examples/controlnet/train_controlnet.py
@@ -1048,7 +1048,8 @@ def main(args):
 
                 # Add noise to the latents according to the noise magnitude at each timestep
                 # (this is the forward diffusion process)
-                noisy_latents = noise_scheduler.add_noise(latents, noise, timesteps)
+                noisy_latents = noise_scheduler.add_noise(
+                    latents.float(), noise.float(), timesteps).to(dtype=weight_dtype)
 
                 # Get the text embedding for conditioning
                 encoder_hidden_states = text_encoder(batch["input_ids"], return_dict=False)[0]
